UserActivity/requirements_personel.txt
appdirs==1.4.3
asgiref==3.2.3
bson==0.5.10
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.4.0
django-crispy-forms==1.9.0
django-elasticsearch-dsl==7.1.1
django-extensions==2.2.8
django-filters==0.2.1
django-geolocation-fields==0.1.8
django-geoposition==0.3.0
django-heroku==0.3.1
django-jsonfield==1.4.0
django-phone-field==1.8.0
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
factory-boy==2.12.0
Faker==4.1.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.0.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==20.0.7
whitenoise==5.1.0
